fix crash on missing comments
Some DMARC reports contain the "comment" field with a null value. This
would case a crash when trying to combine the comments into one.

Therefore, handle None comments as well.

Also remove a buggy line; the reason object already exists in the list,
so no need to append it (and even if we needed to append it, we would
need to iterate over a copy, otherwise we might end up in an endless
loop)